JDBC-User/src/se/mebe/jdbcuser/model/WorkItem.java
package se.mebe.jdbcuser.model;

import java.io.Serializable;

public final class WorkItem implements Serializable {

	private static final long serialVersionUID = -7744087716159944260L;
	private long id;
	private String title;
	private String description;
	private String state;
	private long teamId;

	public WorkItem(long id, String title, String description, String state, long teamId) {
		this.id = id;
		this.title = title;
		this.description = description;
		this.state = state;
		this.teamId = teamId;
	}

	public long getId() {
		return id;
	}

	public void setId(long id) {
		this.id = id;
	}

	public String getTitle() {
		return title;
	}

	public void setTitle(String title) {
		this.title = title;
	}

	public String getDescription() {
		return description;
	}

	public void setDescription(String description) {
		this.description = description;
	}

	public String getState() {
		return state;
	}

	public void setState(String state) {
		this.state = state;
	}

	public long getTeamId() {
		return teamId;
	}

	public void setTeamId(long teamId) {
		this.teamId = teamId;
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((description == null) ? 0 : description.hashCode());
		result = prime * result + (int) (id ^ (id >>> 32));
		result = prime * result + ((state == null) ? 0 : state.hashCode());
		result = prime * result + (int) (teamId ^ (teamId >>> 32));
		result = prime * result + ((title == null) ? 0 : title.hashCode());
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		WorkItem other = (WorkItem) obj;
		if (description == null) {
			if (other.description != null)
				return false;
		} else if (!description.equals(other.description))
			return false;
		if (id != other.id)
			return false;
		if (state == null) {
			if (other.state != null)
				return false;
		} else if (!state.equals(other.state))
			return false;
		if (teamId != other.teamId)
			return false;
		if (title == null) {
			if (other.title != null)
				return false;
		} else if (!title.equals(other.title))
			return false;
		return true;
	}

	@Override
	public String toString() {
		return "WorkItem id = " + id + "\t" + "title = " + title + "\t" + "description = " + description + "\t" + "state = " + state
				+"\t" + "teamId = " + teamId + "\n";
	}

}
